# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StayBuddy API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
...
    "/admin/stays/{stayId}/claims": {
      "post": {
        "operationId": "issueStayClaim",
        "parameters": [
          {
            "name": "stayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueAccess"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "IssuedAccess",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedAccess"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/stays/{stayId}/prearrival-invitations": {
      "post": {
        "operationId": "issuePrearrivalInvitation",
        "parameters": [
          {
            "name": "stayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "IssuedAccess",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedAccess"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/stay-claims/scan": {
      "post": {
        "operationId": "scanStayClaim",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ClaimSession",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimSession"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/prearrival-invitations/scan": {
      "post": {
        "operationId": "scanPrearrivalInvitation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ClaimSession",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimSession"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/email/start": {
      "post": {
        "operationId": "startEmailOtp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOtpStart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OtpChallenge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallenge"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/email/verify": {
      "post": {
        "operationId": "verifyEmailOtp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOtpVerify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GuestSession",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSession"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/auth/oauth": {
      "post": {
        "operationId": "verifyOAuth",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthIdentityAssertion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GuestSession",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSession"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/consents": {
      "post": {
        "operationId": "recordConsent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ConsentReceipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentReceipt"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/stay-claims/complete": {
      "post": {
        "operationId": "completeStayClaim",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCompleteCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "StayActivation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StayActivation"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/prearrival-invitations/complete": {
      "post": {
        "operationId": "completePrearrivalInvitation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCompleteCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "StayActivation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StayActivation"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/me/devices/push-permission": {
      "post": {
        "operationId": "updatePushPermission",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushPermissionCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PushPermissionCommand",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushPermissionCommand"
                }
              }
            }
          },
          "400": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-App-Installation-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "TENANT_NOT_RESOLVED",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "CLAIM_EXPIRED",
              "CLAIM_REPLAYED",
              "CLAIM_REVOKED",
              "OTP_INVALID",
              "OTP_EXPIRED",
              "TERMS_REQUIRED",
              "INTERNAL_ERROR"
            ]
          },
          "traceId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "retryable": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": [
                "string",
                "number",
                "boolean"
              ]
            }
          }
        },
        "required": [
          "code",
          "traceId",
          "retryable"
        ],
        "additionalProperties": false
      },
...
      "IssueAccess": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "roomNumber": {
            "type": "string",
            "maxLength": 40
          },
          "ttlMinutes": {
            "default": 30,
            "type": "integer",
            "minimum": 5,
            "maximum": 120
          }
        },
        "required": [
          "ttlMinutes"
        ],
        "additionalProperties": false
      },
      "IssuedAccess": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "opaqueToken": {
            "type": "string",
            "minLength": 32
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "opaqueToken",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "ClaimScanRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "opaqueToken": {
            "type": "string",
            "minLength": 32,
            "maxLength": 512
          }
        },
        "required": [
          "opaqueToken"
        ],
        "additionalProperties": false
      },
      "ClaimSession": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "claimSessionId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "preview": {
            "type": "object",
            "properties": {
              "hotelDisplayName": {
                "type": "string"
              },
              "checkInDate": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "checkOutDate": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "guestNameMasked": {
                "type": "string"
              }
            },
            "required": [
              "hotelDisplayName",
              "checkInDate",
              "checkOutDate",
              "guestNameMasked"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "claimSessionId",
          "expiresAt",
          "preview"
        ],
        "additionalProperties": false
      },
...
      "EmailOtpStart": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "installationId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "email",
          "installationId"
        ],
        "additionalProperties": false
      },
      "EmailOtpVerify": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          },
          "installationId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "challengeId",
          "code",
          "installationId"
        ],
        "additionalProperties": false
      },
      "OtpChallenge": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "challengeId",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "OAuthIdentityAssertion": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "apple",
              "google"
            ]
          },
          "idToken": {
            "type": "string",
            "minLength": 20
          },
          "installationId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "provider",
          "idToken",
          "installationId"
        ],
        "additionalProperties": false
      },
      "GuestSession": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "hotelId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "sessionId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "accessToken": {
            "type": "string",
            "minLength": 20
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "accountId",
          "hotelId",
          "sessionId",
          "accessToken",
          "expiresAt"
        ],
        "additionalProperties": false
      },
      "ConsentCommand": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "TERMS",
              "PRIVACY",
              "MARKETING",
              "PARTNER_OFFERS"
            ]
          },
          "channel": {
            "type": "string",
            "enum": [
              "SERVICE",
              "EMAIL",
              "PUSH",
              "IN_APP"
            ]
          },
          "granted": {
            "type": "boolean"
          },
          "definitionVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "locale": {
            "type": "string",
            "enum": [
              "en",
              "th",
              "zh-CN",
              "ru"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "ONBOARDING",
              "SETTINGS",
              "POLICY_UPDATE"
            ]
          }
        },
        "required": [
          "purpose",
          "channel",
          "granted",
          "definitionVersion",
          "locale",
          "source"
        ],
        "additionalProperties": false
      },
      "ConsentReceipt": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "TERMS",
              "PRIVACY",
              "MARKETING",
              "PARTNER_OFFERS"
            ]
          },
          "channel": {
            "type": "string",
            "enum": [
              "SERVICE",
              "EMAIL",
              "PUSH",
              "IN_APP"
            ]
          },
          "granted": {
            "type": "boolean"
          },
          "definitionVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "locale": {
            "type": "string",
            "enum": [
              "en",
              "th",
              "zh-CN",
              "ru"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "ONBOARDING",
              "SETTINGS",
              "POLICY_UPDATE"
            ]
          },
          "consentEventId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "purpose",
          "channel",
          "granted",
          "definitionVersion",
          "locale",
          "source",
          "consentEventId"
        ],
        "additionalProperties": false
      },
      "ClaimCompleteCommand": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "claimSessionId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "accountId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 80
          },
          "acceptedTermsVersion": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "claimSessionId",
          "accountId",
          "acceptedTermsVersion"
        ],
        "additionalProperties": false
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}